pyproject.toml
[tool.poetry]
name = "zmet"
version = "0.0.0"
description = "personal knowledge database based on google keep"
authors = ["Tomas Krnak <tomas@krnak.cz>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
beautifulsoup4 = "^4.11.1"
Flask-WTF = "^1.0.1"
Flask-Login = "^0.6.2"
requests = "^2.28.1"
Markdown = "^3.4.1"
gkeepapi = "^0.14.2"
gunicorn = "^20.1.0"
PyYAML = "^6.0"
pycryptodome = "^3.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
